requirements.txt
opencv-python>=4.8.1
Pillow>=10.0.0
